Extract cookie related code to CookieMixin
